# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vendus REST API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/partners": {
            "get": {
                "tags": [
                    "Partners"
                ],
                "summary": "Get list of partners",
                "description": "Returns a paginated list of partners with optional filtering based on various parameters.",
                "operationId": "getPartners",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filter by partner name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Filter by exact partner email",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "commission",
                        "in": "query",
                        "description": "Filter by commission rate",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Partner"
                                            }
                                        },
                                        "links": {
                                            "properties": {
                                                "first": {
                                                    "type": "string",
                                                    "example": "http://localhost:8000/api/partners?page=1"
                                                },
                                                "last": {
                                                    "type": "string",
                                                    "example": "http://localhost:8000/api/partners?page=10"
                                                },
                                                "prev": {
                                                    "type": "string",
                                                    "example": "http://localhost:8000/api/partners?page=1",
                                                    "nullable": true
                                                },
                                                "next": {
                                                    "type": "string",
                                                    "example": "http://localhost:8000/api/partners?page=3",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "from": {
                                                    "type": "integer",
                                                    "example": 16
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "path": {
                                                    "type": "string",
                                                    "example": "http://localhost:8000/api/partners"
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "to": {
                                                    "type": "integer",
                                                    "example": 30
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 150
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Partners"
                ],
                "summary": "Create a new partner",
                "description": "Creates a new partner record in the database.",
                "operationId": "addPartner",
                "requestBody": {
                    "description": "Partner data to create",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Partner"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Partner created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "name": {
                                                    "description": "Full name of the partner",
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "description": "Email address of the partner",
                                                    "type": "string",
                                                    "format": "email"
                                                },
                                                "commission": {
                                                    "description": "Commission rate for the partner",
                                                    "type": "number",
                                                    "format": "float"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error messsage"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": "Object with all the errors by field"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/partners/{id}": {
            "get": {
                "tags": [
                    "Partners"
                ],
                "summary": "Get a single partner",
                "description": "Returns a single partner by ID.",
                "operationId": "getPartnerById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the partner to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "name": {
                                                    "description": "Full name of the partner",
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "description": "Email address of the partner",
                                                    "type": "string",
                                                    "format": "email"
                                                },
                                                "commission": {
                                                    "description": "Commission rate for the partner",
                                                    "type": "number",
                                                    "format": "float"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Partner not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Object not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Partners"
                ],
                "summary": "Update a partner",
                "description": "Updates a partner's record in the database and returns the updated partner.",
                "operationId": "updatePartner",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the partner to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Partner data to update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Partner"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Partner updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "name": {
                                                    "description": "Full name of the partner",
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "description": "Email address of the partner",
                                                    "type": "string",
                                                    "format": "email"
                                                },
                                                "commission": {
                                                    "description": "Commission rate for the partner",
                                                    "type": "number",
                                                    "format": "float"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Partner not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Object not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error messsage"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": "Object with all the errors by field"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Partners"
                ],
                "summary": "Delete a partner",
                "description": "Deletes a specific partner by ID and returns no content.",
                "operationId": "deletePartner",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the partner to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content (successful deletion)"
                    },
                    "404": {
                        "description": "Partner not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Object not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Partner": {
                "title": "Partner",
                "description": "Schema for Partner data",
                "required": [
                    "name",
                    "vat",
                    "email",
                    "address",
                    "commission"
                ],
                "properties": {
                    "id": {
                        "description": "The unique identifier of the partner",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "Full name of the partner",
                        "type": "string"
                    },
                    "vat": {
                        "description": "VAT number of the partner",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email address of the partner",
                        "type": "string",
                        "format": "email"
                    },
                    "address": {
                        "description": "Physical address of the partner",
                        "type": "string"
                    },
                    "commission": {
                        "description": "Commission rate for the partner",
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            },
...
        }
    }
}